Split lifecycle module and clear SonarCloud new-code findings
Split the lifecycle module along the policy/runtime seam to stay under
the 500-line file limit: pure model + evaluators + state stay in
core/lifecycle_policy.py; the enforcement runtime (lock, activity
signal, apply, enforce_once/run_monitor) moves to
core/lifecycle_enforce.py. CLI lifecycle commands move to
cli/lifecycle.py, keeping lab.py under the limit. Add docstrings to all
helpers, type the backend parameters, use collections.abc.Callable,
parametrize generics, and relocate trailing comments in config.py.
